chore: removes storybook-base
