--- .storybook-base/addons.js
@@ -1,7 +1,0 @@
-import 'storybook-readme/register';
-import '@storybook/addon-viewport';
-import '@storybook/addon-a11y/register';
-import '@storybook/addon-options/register';
-import '@storybook/addon-actions/register';
-import '@storybook/addon-links/register';
-import '@storybook/addon-knobs/register';
