--- .storybook-base/loadStories.js
@@ -1,18 +1,0 @@
-/*
-Copyright (c) 2018 Uber Technologies, Inc.
-
-This source code is licensed under the MIT license found in the
-LICENSE file in the root directory of this source tree.
-*/
-
-// automatically import all files ending in *.stories.js
-const req = require.context(
-  '../src',
-  true,
-  /^((?!template-component\/).)*.stories.js$/,
-);
-
-export default function loadStories() {
-  require('../docs/pages/pages.js');
-  req.keys().forEach(filename => req(filename));
-}
