--- .storybook-base/webpack.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  optimization: {
-    minimize: false,
-  },
-};
