+++ .storybook/addons.js
@@ -1,1 +1,7 @@
-import '../.storybook-base/addons.js';
+import 'storybook-readme/register';
+import '@storybook/addon-viewport';
+import '@storybook/addon-a11y/register';
+import '@storybook/addon-options/register';
+import '@storybook/addon-actions/register';
+import '@storybook/addon-links/register';
+import '@storybook/addon-knobs/register';
